+++ src/model/controller.model.ts
@@ -6,16 +6,17 @@ function getElemData(t: any, d: string) {
   return t.getAttribute(ShopCartConst.PREFIX_DATA + d);
 }
 
-module ShopCart {
+namespace ShopCart {
   export var variantProductMap = [];
   export var productVariantsImageMap = [];
   export var variantMap = [];
   export class CartController {
-    config: ShopCart.Config
-    cart: ShopCart.Cart
-    storage: any
-    events: any
-    utm: Array<ShopCart.Utm>
+    config: ShopCart.Config;
+    cart: ShopCart.Cart;
+    storage: any;
+    events: any;
+    utm: Array<ShopCart.Utm>;
+    dispatcher: ShopCart.Dispatcher;
 
     // Public function
 
@@ -29,7 +30,7 @@ module ShopCart {
       var p = ShopCart.variantProductMap[upc];
       var v = ShopCart.variantMap[upc];
       if (!upc || !p || !v) {
-        return null
+        return null;
       }
       return new ShopCart.Item(upc, p, v);
     }
@@ -47,6 +48,7 @@ module ShopCart {
         return null;
       }
       var q = this.cart.setItem(i);
+      this.dispatcher.add(upc);
       this.openCart();
       this.setTotal();
       return q;
@@ -65,6 +67,7 @@ module ShopCart {
         return null;
       }
       var q = this.cart.removeItem(i);
+      this.dispatcher.remove(upc);
       this.setTotal();
       return q;
     }
@@ -76,10 +79,10 @@ module ShopCart {
     */
     getCheckoutUrl(): string {
       var storeUrl = this.config.store.url;
-      if (storeUrl[storeUrl.length - 1] != '/') {
-        storeUrl += '/';
+      if (storeUrl[storeUrl.length - 1] != "/") {
+        storeUrl += "/";
       }
-      var url = storeUrl + 'cart/';
+      var url = storeUrl + "cart/";
       if (this.cart.quantity <= 0) {
         return null;
       }
@@ -87,13 +90,14 @@ module ShopCart {
         if (this.cart.items[i].quantity <= 0) {
           continue;
         }
-        url += (this.cart.items[i].item.upc + ':' + this.cart.items[i].quantity + ',');
+        url +=
+          this.cart.items[i].item.upc + ":" + this.cart.items[i].quantity + ",";
       }
       if (this.utm && this.utm.length > 0) {
-        url += '?';
-        this.utm.forEach(function (u: ShopCart.Utm) {
-          url += (u.key + '=' + u.value + '&');
-        })
+        url += "?";
+        this.utm.forEach(function(u: ShopCart.Utm) {
+          url += u.key + "=" + u.value + "&";
+        });
       }
       return url;
     }
@@ -105,20 +109,29 @@ module ShopCart {
     */
     openCart() {
       this.updateCartQuantities();
-      this.elementLoop(this.config.selector.cart.main, function (e: Element, self: ShopCart.CartController) {
+      this.elementLoop(this.config.selector.cart.main, function(
+        e: Element,
+        self: ShopCart.CartController
+      ) {
         if (e.className.indexOf(self.config.selectorClass.cart.open) >= 0) {
           return;
         }
-        e.className += (' ' + self.config.selectorClass.cart.open);
-        var h = '';
+        e.className += " " + self.config.selectorClass.cart.open;
+        var h = "";
         for (var i = 0; i < self.cart.items.length; i++) {
           if (self.cart.items[i].quantity > 0) {
-            var n = self.createCartItem(self.cart.items[i])
+            var n = self.createCartItem(self.cart.items[i]);
           }
         }
-      })
-      if (document.body.className.indexOf(this.config.selectorClass.body.cartOpen) < 0) {
-        document.body.className += (' ' + this.config.selectorClass.body.cartOpen);
+      });
+      if (
+        document.body.className.indexOf(
+          this.config.selectorClass.body.cartOpen
+        ) < 0
+      ) {
+        document.body.className +=
+          " " + this.config.selectorClass.body.cartOpen;
+        this.dispatcher.open();
       }
     }
 
@@ -142,16 +155,20 @@ module ShopCart {
 
     Registers the events associated to each item that has been added to the cart, in order to perform actions such as increment, decrement and update item quantity.
     */
-    private registerCartItemEvents(n: Element, ) {
-      var l = n.querySelectorAll(ShopCartConst.SELECTOR_CART_ACTION)
+    private registerCartItemEvents(n: Element) {
+      var l = n.querySelectorAll(ShopCartConst.SELECTOR_CART_ACTION);
       for (var i = 0; i < l.length; i++) {
-        this.registerCartActionEvent(l[i])
-        this.events[getElemData(l[i], ShopCartConst.DATA_UPC) + '-' + getElemData(l[i], ShopCartConst.DATA_ACTION)] = true;
+        this.registerCartActionEvent(l[i]);
+        this.events[
+          getElemData(l[i], ShopCartConst.DATA_UPC) +
+            "-" +
+            getElemData(l[i], ShopCartConst.DATA_ACTION)
+        ] = true;
       }
 
-      l = n.querySelectorAll('input' + ShopCartConst.SELECTOR_ITEM_QUANTITY)
+      l = n.querySelectorAll("input" + ShopCartConst.SELECTOR_ITEM_QUANTITY);
       for (var i = 0; i < l.length; i++) {
-        this.registerCartUpdateEvent(l[i])
+        this.registerCartUpdateEvent(l[i]);
       }
     }
 
@@ -161,15 +178,18 @@ module ShopCart {
     Registers the update quantity events associated to each item that has been added to the cart.
     */
     private registerCartUpdateEvent(el: Element) {
-      var self = this
-      el.addEventListener('input', function (e) {
+      var self = this;
+      el.addEventListener("input", function(e) {
         var v = (<any>e.target).value;
         if (v) {
           var q = parseInt(v);
           if (isNaN(q)) {
             return;
           }
-          self.setItemQuantity(getElemData(e.target, ShopCartConst.DATA_UPC), q);
+          self.setItemQuantity(
+            getElemData(e.target, ShopCartConst.DATA_UPC),
+            q
+          );
         }
       });
     }
@@ -182,25 +202,29 @@ module ShopCart {
     private registerCartActionEvent(el: Element) {
       var self = this;
       if (!this.events) {
-        this.events = {}
+        this.events = {};
       }
       if (this.events[getElemData(el, ShopCartConst.DATA_UPC)]) {
         return;
       }
-      this.events[getElemData(el, ShopCartConst.DATA_UPC) + '-' + getElemData(el, ShopCartConst.DATA_ACTION)] = true;
-      el.addEventListener('click', function (e) {
+      this.events[
+        getElemData(el, ShopCartConst.DATA_UPC) +
+          "-" +
+          getElemData(el, ShopCartConst.DATA_ACTION)
+      ] = true;
+      el.addEventListener("click", function(e) {
         var upc = getElemData(e.target, ShopCartConst.DATA_UPC);
         switch (getElemData(e.target, ShopCartConst.DATA_ACTION)) {
           case ShopCartConst.ACTION_ADD:
             var q = self.setItem(upc);
-            self.updateCartQuantities()
+            self.updateCartQuantities();
             break;
           case ShopCartConst.ACTION_REMOVE:
             var q = self.removeItem(upc);
-            self.updateCartQuantities()
-            break
+            self.updateCartQuantities();
+            break;
         }
-      })
+      });
     }
 
     /*
@@ -214,18 +238,18 @@ module ShopCart {
     private setItemQuantity(upc: number, quantity: number): number {
       var i = this.getItem(upc);
       if (!i) {
-        return null
+        return null;
       }
       var exists = false;
       var q = 0;
-      this.cart.items.map(function (ic) {
+      this.cart.items.map(function(ic) {
         if (ic.item.upc == upc) {
           ic.quantity = quantity;
           exists = true;
           q = ic.quantity;
         }
         return ic;
-      })
+      });
       if (!exists) {
         this.cart.items.push(new ShopCart.CartItem(i, quantity));
         q = 1;
@@ -242,13 +266,18 @@ module ShopCart {
     If the element already exists, updates the quantity.
     */
     private createCartItem(i: CartItem) {
-      if (i.quantity > 0 && !document.getElementById(ShopCartConst.PREFIX_ITEM + i.item.upc)) {
-        var n = document.createElement('div');
+      if (
+        i.quantity > 0 &&
+        !document.getElementById(ShopCartConst.PREFIX_ITEM + i.item.upc)
+      ) {
+        var n = document.createElement("div");
         n.className = ShopCartConst.CLASS_ITEM;
         n.id = ShopCartConst.PREFIX_ITEM + i.item.upc;
         n.innerHTML = i.getTemplate(this.config);
-        document.querySelector(ShopCartConst.SELECTOR_CART_ITEM_LIST).appendChild(n);
-        this.registerCartItemEvents(n)
+        document
+          .querySelector(ShopCartConst.SELECTOR_CART_ITEM_LIST)
+          .appendChild(n);
+        this.registerCartItemEvents(n);
       } else {
         this.updateItemQuantity(i);
       }
@@ -272,11 +301,21 @@ module ShopCart {
     For each HTML element associated to the given cart item, update the displayed quantity.
     */
     private updateItemQuantity(i: ShopCart.CartItem) {
-      var n = document.querySelectorAll('#' + ShopCartConst.PREFIX_ITEM + i.item.upc),
-        q = document.querySelectorAll('#' + ShopCartConst.PREFIX_ITEM + i.item.upc + ' ' + ShopCartConst.SELECTOR_ITEM_QUANTITY);
+      var n = document.querySelectorAll(
+          "#" + ShopCartConst.PREFIX_ITEM + i.item.upc
+        ),
+        q = document.querySelectorAll(
+          "#" +
+            ShopCartConst.PREFIX_ITEM +
+            i.item.upc +
+            " " +
+            ShopCartConst.SELECTOR_ITEM_QUANTITY
+        );
       if (q && q.length > 0) {
         if (i.quantity <= 0) {
-          document.querySelector(ShopCartConst.SELECTOR_CART_ITEM_LIST).removeChild(n[0]);
+          document
+            .querySelector(ShopCartConst.SELECTOR_CART_ITEM_LIST)
+            .removeChild(n[0]);
           return;
         }
         for (var j = 0; j < q.length; j++) {
@@ -292,10 +331,12 @@ module ShopCart {
     Returns the cart total.
     */
     private setTotal(): number {
-      var t = 0, q = 0;
+      var t = 0,
+        q = 0;
       for (var i = 0; i < this.cart.items.length; i++) {
-        t += (this.cart.items[i].item.product.price * this.cart.items[i].quantity);
-        q += (this.cart.items[i].quantity > 0 ? this.cart.items[i].quantity : 0);
+        t +=
+          this.cart.items[i].item.product.price * this.cart.items[i].quantity;
+        q += this.cart.items[i].quantity > 0 ? this.cart.items[i].quantity : 0;
       }
       this.cart.total = t;
       this.cart.quantity = q;
@@ -334,19 +375,19 @@ module ShopCart {
     */
     private getUrlUtm(): Array<ShopCart.Utm> {
       var utm = [];
-      var utmPrefix = 'utm_';
+      var utmPrefix = "utm_";
       var url = window.location.href.trim();
-      if (url.indexOf('?') < 0 || url.indexOf(utmPrefix) < 0) {
+      if (url.indexOf("?") < 0 || url.indexOf(utmPrefix) < 0) {
         return utm;
       }
-      var q = url.split('?')[1];
+      var q = url.split("?")[1];
       var regex = new RegExp("^" + utmPrefix + "?[a-zA-Z0-9]*");
-      q.split('&').forEach(function (p) {
+      q.split("&").forEach(function(p) {
         if (regex.test(p)) {
-          var i = p.split('=');
+          var i = p.split("=");
           utm.push(new ShopCart.Utm(i[0], i[1]));
         }
-      })
+      });
       return utm;
     }
 
@@ -371,7 +412,8 @@ module ShopCart {
     private updateCartTotal() {
       var t = document.querySelectorAll(this.config.selector.cart.total);
       for (var i = 0; i < t.length; i++) {
-        t[i].innerHTML = this.cart.total.toFixed(2) + ' ' + this.cart.currency.label;
+        t[i].innerHTML =
+          this.cart.total.toFixed(2) + " " + this.cart.currency.label;
       }
       var t = document.querySelectorAll(this.config.selector.cart.count);
       for (var i = 0; i < t.length; i++) {
@@ -385,13 +427,18 @@ module ShopCart {
     Updates cart quantity status in the page.
     */
     private updateCartCheckoutButton() {
-      this.elementLoop(this.config.selector.cart.main, function (e: Element, self: ShopCart.CartController) {
+      this.elementLoop(this.config.selector.cart.main, function(
+        e: Element,
+        self: ShopCart.CartController
+      ) {
         if (self.cart.quantity > 0) {
-          e.className = e.className.replace(self.config.selectorClass.cart.empty, '').trim();
+          e.className = e.className
+            .replace(self.config.selectorClass.cart.empty, "")
+            .trim();
           self.setCartProductClass();
         } else {
           if (e.className.indexOf(self.config.selectorClass.cart.empty) < 0) {
-            e.className += (' ' + self.config.selectorClass.cart.empty);
+            e.className += " " + self.config.selectorClass.cart.empty;
           }
         }
       });
@@ -403,16 +450,19 @@ module ShopCart {
     Updates cart product-related classes in the page.
     */
     private setCartProductClass() {
-      this.elementLoop(this.config.selector.cart.main, function (e: Element, self: ShopCart.CartController) {
+      this.elementLoop(this.config.selector.cart.main, function(
+        e: Element,
+        self: ShopCart.CartController
+      ) {
         for (var j = 0; j < self.config.products.length; j++) {
           var id = self.config.products[j].id;
-          e.className = e.className.replace(id, '').trim();
+          e.className = e.className.replace(id, "").trim();
         }
         for (var j = 0; j < self.cart.items.length; j++) {
           if (self.cart.items[j].quantity > 0) {
             var p = ShopCart.variantProductMap[self.cart.items[j].item.upc];
             if (e.className.indexOf(p.id) < 0) {
-              e.className += (' ' + p.id)
+              e.className += " " + p.id;
             }
           }
         }
@@ -425,15 +475,22 @@ module ShopCart {
     Updates cart status in the page.
     */
     private updateCartActive() {
-      this.elementLoop(this.config.selector.element.main, function (e: Element, self: ShopCart.CartController) {
+      this.elementLoop(this.config.selector.element.main, function(
+        e: Element,
+        self: ShopCart.CartController
+      ) {
         if (self.cart.quantity > 0) {
-          if (e.className.indexOf(self.config.selectorClass.element.active) < 0) {
-            e.className += (' ' + self.config.selectorClass.element.active);
+          if (
+            e.className.indexOf(self.config.selectorClass.element.active) < 0
+          ) {
+            e.className += " " + self.config.selectorClass.element.active;
           }
         } else {
-          e.className = e.className.replace(self.config.selectorClass.element.active, '').trim();
+          e.className = e.className
+            .replace(self.config.selectorClass.element.active, "")
+            .trim();
         }
-      })
+      });
     }
 
     /*
@@ -442,14 +499,21 @@ module ShopCart {
     Updates cart notification status and quantity in the page.
     */
     private updateCartNotification() {
-      this.elementLoop(this.config.selector.element.notification, function (e: Element, self: ShopCart.CartController) {
+      this.elementLoop(this.config.selector.element.notification, function(
+        e: Element,
+        self: ShopCart.CartController
+      ) {
         if (self.cart.quantity > 0) {
-          e.innerHTML = self.cart.quantity.toString()
-          if (e.className.indexOf(self.config.selectorClass.element.active) < 0) {
-            e.className += (' ' + self.config.selectorClass.element.active);
+          e.innerHTML = self.cart.quantity.toString();
+          if (
+            e.className.indexOf(self.config.selectorClass.element.active) < 0
+          ) {
+            e.className += " " + self.config.selectorClass.element.active;
           }
         } else {
-          e.className = e.className.replace(self.config.selectorClass.element.active, '').trim();
+          e.className = e.className
+            .replace(self.config.selectorClass.element.active, "")
+            .trim();
         }
       });
     }
@@ -460,20 +524,36 @@ module ShopCart {
     Set the body to be cart-ready when ShopCart is ready to work.
     */
     private ready() {
-      return document.body.className += (' ' + this.config.selectorClass.body.cartReady);
+      this.dispatcher.ready();
+      return (document.body.className +=
+        " " + this.config.selectorClass.body.cartReady);
     }
 
     private setConfig(config: any) {
       if (!config.store || !config.store.url) {
-        throw Error("Store url is missing")
+        throw Error("Store url is missing");
       }
       this.config = new ShopCart.Config(config);
       for (var i = 0; i < config.products.length; i++) {
         var p = config.products[i];
         for (var j = 0; j < p.variants.length; j++) {
-          ShopCart.variantProductMap[p.variants[j].upc] = new ShopCart.Product(p.id, p.name, p.price, p.description, p.image.url);
-          ShopCart.productVariantsImageMap[p.variants[j].upc] = new ShopCart.ProductImage(p.variants[j].image.url);
-          ShopCart.variantMap[p.variants[j].upc] = new ProductVariant(p.variants[j].name, p.variants[j].upc, p.variants[j].sku, p.variants[j].size, p.variants[j].image.url);
+          ShopCart.variantProductMap[p.variants[j].upc] = new ShopCart.Product(
+            p.id,
+            p.name,
+            p.price,
+            p.description,
+            p.image.url
+          );
+          ShopCart.productVariantsImageMap[
+            p.variants[j].upc
+          ] = new ShopCart.ProductImage(p.variants[j].image.url);
+          ShopCart.variantMap[p.variants[j].upc] = new ProductVariant(
+            p.variants[j].name,
+            p.variants[j].upc,
+            p.variants[j].sku,
+            p.variants[j].size,
+            p.variants[j].image.url
+          );
         }
       }
     }
@@ -486,15 +566,19 @@ module ShopCart {
         if (!(<any>window).store) {
           throw Error("Storage is missing");
         }
-        storage = (<any>window).store
+        storage = (<any>window).store;
       }
       if (!cart) {
         cart = (<any>window).store.get(ShopCartConst.STORAGE_CART);
       }
+      this.dispatcher = new Dispatcher();
+      this.dispatcher.init();
       this.setConfig(config);
       this.storage = storage;
       this.setUtm();
-      cart = ShopCart.isValidCart(cart) ? new ShopCart.Cart(cart.currency, cart) : new ShopCart.Cart(this.config.currency);
+      cart = ShopCart.isValidCart(cart)
+        ? new ShopCart.Cart(cart.currency, cart)
+        : new ShopCart.Cart(this.config.currency);
       this.cart = cart;
       this.setTotal();
       this.ready();