+++ src/model/utm.model.ts
@@ -1,9 +1,9 @@
 /// <reference path="./index.d.ts" />
 
-module ShopCart {
+namespace ShopCart {
   export class Utm {
-    key: string
-    value: string
+    key: string;
+    value: string;
     constructor(k, v) {
       this.key = k;
       this.value = v;