+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Unreleased
-
+- Introducing cart events
 
 ## 1.1.0 (2018-01-04)
 - New product image configuration